// GCE_COURSE_WEB/src/app/core/models/personne.ts
export class Personne{
    id : string;
    nom: string;
    prenom: string;
    codeNaturePersonne: string;
    idBanque: string;
    codeCategorie: string;
    dateNaissance: Date;
    adresse1: string;
    numeroTelephone1: string;
    email: string;
    rib: string;
    dateCreation: Date;
    codePersonne:string;
    codeSexe: string;
    codeTypePieceIdentite: string;
    numeroPieceIdentite: string;
    databaseSource: string;
    codeTypePersonne: string;
    idEmployeur: string;

    //Morale
    nomGerant: string;
    raisonSociale: string;
    categoriePersMora: string;

    //Asso
    designation: string;
    
}